[559825] Add 3 command lines: Refresh Representations, Remove Hidden Elements and Export All Images
Bug: 559825
