+++ core/plugins/org.polarsys.capella.core.commandline.core.ui/META-INF/MANIFEST.MF
@@ -8,7 +8,9 @@ Require-Bundle: org.eclipse.core.resources,
  org.eclipse.core.runtime,
  org.eclipse.ui.ide,
  org.polarsys.capella.core.commandline.core,
- org.polarsys.capella.common.tools.report
+ org.polarsys.capella.common.tools.report,
+ org.eclipse.swt,
+ org.eclipse.ui.workbench
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Export-Package: org.polarsys.capella.core.commandline.core.ui
