+++ core/plugins/org.polarsys.capella.core.platform.sirius.ui.actions/src/org/polarsys/capella/core/platform/sirius/ui/handlers/messages.properties
@@ -1,0 +1,34 @@
+#===============================================================================
+# Copyright (c) 2006, 2019 THALES GLOBAL SERVICES.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License v1.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+#  
+# Contributors:
+#    Thales - initial API and implementation
+#===============================================================================
+
+RefreshRepresentation_0=Refresh Representation
+RefreshRepresentation_1=Confirm Capella Representation Refresh
+RefreshRepresentation_2={0} representation(s) might be refreshed\nIt's recommended to backup your model before running this command\n\nAre you sure you want to proceed?
+RefreshRepresentation_6=Refresh ''{0}'' diagram ({1}/{2}).
+
+RefreshRepresentation_4=Save Resource Before Refreshing
+RefreshRepresentation_5=Your session has been modified and is in dirty mode. Do you want to save changes and close the related editor(s) to continue refreshing?
+
+RefreshRepresentation_7=Opening diagram
+RefreshRepresentation_8=Closing diagram
+GotoRelatedElementsHandler_name={0}
+
+RefreshRepresentation_9=No Representation To Refresh
+RefreshRepresentation_10=There is not any representation under the selected element.
+
+RemoveHiddenElementsHandler_NoDiagramDialog_Text=No diagram below the selected element.
+RemoveHiddenElementsHandler_JobName=Remove Hidden Elements from Diagrams
+RemoveHiddenElementsHandler_ConfirmRefreshDialog_Title=Remove element from diagrams
+RemoveHiddenElementsHandler_ConfirmRefreshDialog_Text=Are you sure you want to remove hidden elements from diagrams ?
+
+resource_prefix=platform:/resource/
+refreshResultsFileName=refresh-results.html
+removeHiddenElementsResultsFileName=removeHiddenElements-results.html
