+++ core/plugins/org.polarsys.capella.core.platform.sirius.ui.navigator/plugin.properties
@@ -67,12 +67,6 @@ showInCapellaExplorerAction=Show in Project Explorer
 showImpactAnalysis=Show Impact Analysis...
 gotoRelatedElementsAction=Go To
 
-command.refreshDiagrams=Refresh All Sub Representations
-command.refreshDiagrams.description=Open, refresh, fix any filter problems then close all diagrams after a Repair/Migrate
-
-command.removeHiddenElements=Remove Hidden Elements
-command.removeHiddenElements.description=Remove hidden elements from diagrams
-
 NewRepresentationActionProvider_NewRepresentationAction_Title=New Diagram / Table...
 
 providerName = Eclipse.org
