+++ core/plugins/org.polarsys.capella.core.platform.sirius.ui.navigator/plugin.xml
@@ -1140,20 +1140,6 @@
             id="org.polarsys.capella.core.platform.sirius.ui.navigator.showImpactAnalysisCommand"
             name="%showImpactAnalysis">
       </command>
-      <command
-            categoryId="org.polarsys.capella.core.platform.sirius.ui.navigator"
-            defaultHandler="org.polarsys.capella.core.platform.sirius.ui.navigator.handlers.RefreshDiagramsCommandHandler"
-            id="org.polarsys.capella.core.platform.sirius.ui.navigator.refreshAird"
-            name="%command.refreshDiagrams"
-            description="%command.refreshDiagrams.description">
-      </command>
-      <command
-            categoryId="org.polarsys.capella.core.platform.sirius.ui.navigator"
-            defaultHandler="org.polarsys.capella.core.platform.sirius.ui.navigator.handlers.RemoveHiddenElementsHandler"
-            id="org.polarsys.capella.core.platform.sirius.ui.navigator.removeHiddenElements"
-            name="%command.removeHiddenElements"
-            description="%command.removeHiddenElements.description">
-      </command>
       
    </extension>
    
@@ -1163,14 +1149,6 @@
       <image
             commandId="org.polarsys.capella.core.platform.sirius.ui.navigator.showInDiagramCommand"
             icon="icons/showInDiagram.gif">
-      </image>
-      <image
-            commandId="org.polarsys.capella.core.platform.sirius.ui.navigator.refreshAird"
-            icon="icons/viewpoint.gif">
-      </image>
-      <image
-            commandId="org.polarsys.capella.core.platform.sirius.ui.navigator.removeHiddenElements"
-            icon="icons/viewpoint.gif">
       </image>
       <image
             commandId="org.polarsys.capella.core.platform.sirius.ui.navigator.locateInCapellaExplorerCommand"
@@ -1193,76 +1171,6 @@
    <extension
          point="org.eclipse.ui.menus">
          
-      <menuContribution 
-      		locationURI="popup:capella.project.explorer#PopupMenu?after=group.project">
-         <command
-               commandId="org.polarsys.capella.core.platform.sirius.ui.navigator.refreshAird"
-               label="%command.refreshDiagrams"
-               style="push">
-            <visibleWhen checkEnabled="true">
-               <with variable="selection">
-                  <and>
-                     <iterate>
-                        <or>
-                           <and>
-                              <instanceof
-                                    value="org.eclipse.core.resources.IFile">
-                              </instanceof>
-                              <test
-                                    forcePluginActivation="true"
-                                    property="org.eclipse.core.resources.extension"
-                                    value="aird">
-                              </test>
-                           </and>
-                           <instanceof
-                                 value="org.polarsys.capella.common.data.modellingcore.ModelElement">
-                           </instanceof>
-                        </or>
-                     </iterate>
-                     <count value="1">
-                     </count>
-                  </and>
-               </with>
-            </visibleWhen>
-         </command>
-         <command
-               commandId="org.polarsys.capella.core.platform.sirius.ui.navigator.removeHiddenElements"
-               label="%command.removeHiddenElements"
-               style="push">
-            <visibleWhen
-                  checkEnabled="true">
-               <with
-                     variable="selection">
-                  <and>
-                     <iterate>
-                        <or>
-                           <and>
-                              <instanceof
-                                    value="org.eclipse.core.resources.IFile">
-                              </instanceof>
-                              <test
-                                    forcePluginActivation="true"
-                                    property="org.eclipse.core.resources.extension"
-                                    value="aird">
-                              </test>
-                           </and>
-                           <instanceof
-                                 value="org.polarsys.capella.common.data.modellingcore.ModelElement">
-                           </instanceof>
-                        </or>
-                     </iterate>
-                     <count
-                           value="1">
-                     </count>
-                  </and>
-               </with>
-            </visibleWhen>
-         </command>
-      </menuContribution>
-      
-      
-      
-      
       <menuContribution
             locationURI="popup:org.eclipse.ui.popup.any?after=group.showIn">
          <command
@@ -1529,20 +1437,5 @@
             class="org.polarsys.capella.core.platform.sirius.ui.navigator.view.CapellaStatusLineContributionItemProvider">
          <Priority name="High"/>
       </statusLineContributionItemProvider>
-   </extension>
-   <extension
-         point="org.polarsys.capella.core.commandline.core.commandline">
-      <CommandlineExtension
-            class="org.polarsys.capella.core.platform.sirius.ui.navigator.commandline.RefreshAirdCommandLine"
-            id="org.polarsys.capella.refreshRepresentations">
-      </CommandlineExtension>
-      <CommandlineExtension
-            class="org.polarsys.capella.core.platform.sirius.ui.navigator.commandline.RemoveHiddenElementsCommandLine"
-            id="org.polarsys.capella.removeHiddenElements">
-      </CommandlineExtension>
-      <CommandlineExtension
-            class="org.polarsys.capella.core.platform.sirius.ui.navigator.commandline.ExportRepresentationsCommandLine"
-            id="org.polarsys.capella.exportRepresentations">
-      </CommandlineExtension>
    </extension>
 </plugin>
