+++ core/plugins/org.polarsys.capella.core.sirius.analysis/META-INF/MANIFEST.MF
@@ -33,7 +33,6 @@ Require-Bundle: org.eclipse.core.expressions,
  org.polarsys.capella.core.diagram.helpers;visibility:=reexport,
  org.eclipse.core.expressions,
  org.polarsys.capella.common.queries,
- org.polarsys.capella.core.platform.sirius.ui.actions,
  org.polarsys.capella.common.libraries,
  org.eclipse.sirius.diagram.ui,
  com.google.guava,
@@ -41,7 +40,8 @@ Require-Bundle: org.eclipse.core.expressions,
  org.polarsys.capella.core.preferences,
  org.polarsys.capella.core.data.gen.edit.decorators,
  org.polarsys.capella.core.libraries,
- org.polarsys.capella.core.data.gen
+ org.polarsys.capella.core.data.gen,
+ org.polarsys.capella.core.platform.sirius.ui.actions;visibility:=reexport
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Localization: plugin
