+++ core/plugins/org.polarsys.capella.core.sirius.ui/META-INF/MANIFEST.MF
@@ -29,10 +29,14 @@ Require-Bundle: org.eclipse.ui.ide,
  org.polarsys.capella.core.platform.sirius.ui.actions,
  org.polarsys.capella.core.af.integration,
  org.eclipse.sirius.diagram.formatdata,
- org.polarsys.capella.common.tools.report.appenders.usage;visibility:=reexport
+ org.polarsys.capella.common.tools.report.appenders.usage;visibility:=reexport,
+ org.polarsys.capella.core.sirius.analysis,
+ org.polarsys.capella.core.commandline.core,
+ org.polarsys.capella.core.commandline.core.ui
 Export-Package: org.polarsys.capella.core.sirius.ui,
  org.polarsys.capella.core.sirius.ui.actions,
  org.polarsys.capella.core.sirius.ui.closeproject,
+ org.polarsys.capella.core.sirius.ui.commandline,
  org.polarsys.capella.core.sirius.ui.copyformat,
  org.polarsys.capella.core.sirius.ui.copyformat.keyproviders,
  org.polarsys.capella.core.sirius.ui.danalysis,
