+++ core/plugins/org.polarsys.capella.core.sirius.ui/plugin.properties
@@ -15,3 +15,12 @@ providerName = Eclipse.org
 pluginName = Capella Sirius UI Plug-in
 
 Capella_Rename_Participant_name=Capella Rename Participant
+
+command.refreshDiagrams=Refresh All Sub Representations
+command.refreshDiagrams.description=Open, refresh, fix any filter problems then close all diagrams after a Repair/Migrate
+
+command.removeHiddenElements=Remove Hidden Elements
+command.removeHiddenElements.description=Remove hidden elements from diagrams
+
+category.capella.name=Representations (Capella)
+category.capella.description=
