+++ core/plugins/org.polarsys.capella.core.sirius.ui/src/org/polarsys/capella/core/sirius/ui/commandline/ExportRepresentationsArgumentHelper.java
@@ -8,7 +8,7 @@
  * Contributors:
  *    Thales - initial API and implementation
  *******************************************************************************/
-package org.polarsys.capella.core.platform.sirius.ui.navigator.commandline;
+package org.polarsys.capella.core.sirius.ui.commandline;
 
 import org.eclipse.sirius.common.tools.api.resource.ImageFileFormat;
 import org.polarsys.capella.core.commandline.core.CommandLineArgumentHelper;
