+++ core/plugins/org.polarsys.capella.core.sirius.ui/src/org/polarsys/capella/core/sirius/ui/commandline/messages.properties
@@ -1,0 +1,15 @@
+#===============================================================================
+# Copyright (c) 2006, 2019 THALES GLOBAL SERVICES.
+# All rights reserved. This program and the accompanying materials
+# are made available under the terms of the Eclipse Public License v1.0
+# which accompanies this distribution, and is available at
+# http://www.eclipse.org/legal/epl-v10.html
+#  
+# Contributors:
+#    Thales - initial API and implementation
+#===============================================================================
+
+RefreshRepresentation_0=Refresh Representation
+resource_prefix=platform:/resource/
+refreshResultsFileName=refresh-results.html
+removeHiddenElementsResultsFileName=removeHiddenElements-results.html
